feat(axis): ✨ Update axis data types to int16_t for improved precision
* Changed axis-related data types from `int32_t` to `int16_t` across multiple files.
* Updated function signatures for `setAxisCustomCurve` and `setSize` to reflect new data types.
* Adjusted `ANALOG_AXIS_COUNT` to 6 for expanded axis support.

+++ src/AnalogAxis.cpp
@@ -78,7 +78,7 @@ void AnalogAxisManager::setAxisResponseCurve(uint8_t axis, ResponseCurveType typ
     if (axis < ANALOG_AXIS_COUNT) _curves[axis].setType(type);
 }
 
-void AnalogAxisManager::setAxisCustomCurve(uint8_t axis, const int32_t* table, uint8_t points) {
+void AnalogAxisManager::setAxisCustomCurve(uint8_t axis, const int16_t* table, uint8_t points) {
     if (axis < ANALOG_AXIS_COUNT) _curves[axis].setCustomCurve(table, points);
 }
